Integrate Northern Sami alphanumeric words
