+++ nemo_text_processing/text_normalization/se/taggers/word.py
@@ -14,7 +14,16 @@
 import pynini
 from pynini.lib import pynutil
 
-from nemo_text_processing.text_normalization.en.graph_utils import NEMO_NOT_SPACE, GraphFst
+from nemo_text_processing.text_normalization.en.graph_utils import (
+    NEMO_DIGIT,
+    NEMO_NOT_SPACE,
+    NEMO_SIGMA,
+    GraphFst,
+    convert_space,
+)
+from nemo_text_processing.text_normalization.se.graph_utils import SE_UPPER
+from nemo_text_processing.text_normalization.se.taggers.cardinal import CardinalFst
+from nemo_text_processing.text_normalization.se.utils import get_abs_path
 
 
 class WordFst(GraphFst):
@@ -23,11 +32,43 @@ class WordFst(GraphFst):
         e.g. hund -> tokens { name: "hund" }
 
     Args:
+        cardinal: cardinal graph used to verbalize digit runs in mixed strings
         deterministic: if True will provide a single transduction option,
             for False multiple transduction are generated (used for audio-based normalization)
     """
 
-    def __init__(self, deterministic: bool = True):
-        super().__init__(name="word", kind="classify")
+    def __init__(self, cardinal: CardinalFst, deterministic: bool = True):
+        super().__init__(name="word", kind="classify", deterministic=deterministic)
         word = pynutil.insert("name: \"") + pynini.closure(NEMO_NOT_SPACE, 1) + pynutil.insert("\"")
         self.fst = word.optimize()
+
+        digit = pynini.invert(pynini.string_file(get_abs_path("data/numbers/digit.tsv")))
+        zero = pynini.invert(pynini.string_file(get_abs_path("data/numbers/zero.tsv")))
+        single_digit = digit | zero
+        space = pynutil.insert(" ")
+
+        leading_zero = ("0" + pynini.closure(NEMO_DIGIT, 1)) @ cardinal.graph_with_leading_zero
+        up_to_three = ((NEMO_DIGIT - "0") + pynini.closure(NEMO_DIGIT, 0, 2)) @ cardinal.graph_with_leading_zero
+        four_digits = (
+            ((NEMO_DIGIT - "0") + NEMO_DIGIT) @ cardinal.graph_with_leading_zero
+            + space
+            + ((NEMO_DIGIT**2) @ cardinal.graph_with_leading_zero)
+        )
+        long_digits = single_digit + pynini.closure(space + single_digit, 4)
+        digits = leading_zero | up_to_three | four_digits | long_digits
+        letters = SE_UPPER + pynini.closure(space + SE_UPPER)
+
+        alphanumeric = pynini.union(
+            letters
+            + space
+            + digits
+            + pynini.closure(space + letters + space + digits)
+            + pynini.closure(space + letters, 0, 1),
+            digits
+            + space
+            + letters
+            + pynini.closure(space + digits + space + letters)
+            + pynini.closure(space + digits, 0, 1),
+        )
+        alphanumeric = pynini.cdrewrite(pynutil.delete("-"), "", "", NEMO_SIGMA) @ alphanumeric
+        self.alphanumeric = pynutil.insert('name: "') + convert_space(alphanumeric) + pynutil.insert('"')
